Committing the changes made to create coverage and simplify files and add .gitignore
I wanted to create commit this since testing with basic unit tests
worked well, and I moved all helpers and everything into
DiningPhilosophers.c. I wanted to start on mocks, but I think my focus
for this assessment should be around the testing that is specified
within the req.

This is all working, and I will work on python-based functional testing.

I will then work on my "ultima thule" version on a separate branch when
everything is done-done on main.