UltimaThule: fixes based on Valgrind checks (#9)
Ran valgrind memcheck, drd, and helgrind.

--memcheck is and was good
--drd(data race detection) was noting a lot of "errors", but it is just
because it can access data from other threads. As written, we shouldn't
be able to write
--helgrind I switched to following the guideline of having a global ids
that just follow the thread ids, and the lowest values are grabbed
before the next highest.

Removed unnecessary frees in main, even though free(NULL) is ok, no need
to run it.